+++ src/node-project.ts
@@ -17,6 +17,21 @@ import { Version } from './version';
 
 const PROJEN_SCRIPT = 'projen';
 
+/**
+ * The node package manager to use.
+ */
+export enum NodePackageManager {
+  /**
+   * Use `yarn` as the package manager.
+   */
+  YARN,
+
+  /**
+   * Use `npm` as the package manager.
+   */
+  NPM
+}
+
 export interface NodeProjectCommonOptions {
   readonly bundledDependencies?: string[];
   readonly dependencies?: Record<string, Semver>;
@@ -155,6 +170,13 @@ export interface NodeProjectCommonOptions {
    * @default "registry.npmjs.org"
    */
   readonly npmRegistry?: string;
+
+  /**
+   * The Node Package Manager used to execute scripts
+   *
+   * @default packageManager.YARN
+   */
+  readonly packageManager?: NodePackageManager;
 
   /**
    * License copyright owner.
@@ -454,6 +476,16 @@ export class NodeProject extends Project {
 
   protected readonly npmRegistry: string;
 
+  /**
+   * The package manager to use.
+   */
+  protected readonly packageManager: NodePackageManager;
+
+  /**
+   * The command to use to run scripts (e.g. `yarn run` or `npm run` depends on the package mabnager).
+   */
+  public readonly runScriptCommand: string;
+
   constructor(options: NodeProjectOptions) {
     super();
 
@@ -480,6 +512,16 @@ export class NodeProject extends Project {
     this.npmDistTag = options.npmDistTag ?? 'latest';
 
     this.npmRegistry = options.npmRegistry ?? 'registry.npmjs.org';
+
+    this.packageManager = options.packageManager ?? NodePackageManager.YARN;
+
+    this.runScriptCommand = (() => {
+      switch (this.packageManager) {
+        case NodePackageManager.NPM: return 'npm run';
+        case NodePackageManager.YARN: return 'yarn run';
+        default: throw new Error(`unexpected package manager ${this.packageManager}`);
+      }
+    })();
 
     this.scripts = {};
 
@@ -823,7 +865,7 @@ export class NodeProject extends Project {
   /**
    * Replaces the contents of an npm package.json script.
    *
-   * @param name The script namne
+   * @param name The script name
    * @param commands The commands to run (joined by "&&")
    */
   public addScript(name: string, ...commands: string[]) {
@@ -1012,8 +1054,6 @@ export class NodeProject extends Project {
   public postSynthesize(outdir: string) {
     super.postSynthesize(outdir);
 
-    const install = ['yarn install'];
-
     // now we run `yarn install`, but before we do that, remove the
     // `node_modules/projen` symlink so that yarn won't hate us.
     const projenModule = path.resolve('node_modules', 'projen');
@@ -1023,18 +1063,28 @@ export class NodeProject extends Project {
       }
     } catch (e) { }
 
-    // add --check-files to ensure all modules exist (especiall projen which was just removed).
-    install.push('--check-files');
-
-    // if we are running in a CI environment, fix versions through the lockfile.
-    if (process.env.CI) {
-      logging.info('Running yarn with --frozen-lockfile since "CI" is defined.');
-      install.push('--frozen-lockfile');
-    }
-
-    exec(install.join(' '), { cwd: outdir });
+    exec(this.installDepsCommand, { cwd: outdir });
 
     this.resolveDependencies(outdir);
+  }
+
+  private get installDepsCommand() {
+    switch (this.packageManager) {
+      case NodePackageManager.YARN:
+        return [
+          'yarn install',
+          '--check-files', // ensure all modules exist (especially projen which was just removed).
+          ...process.env.CI ? ['--frozen-lockfile'] : [],
+        ].join(' ');
+
+      case NodePackageManager.NPM:
+        return process.env.CI
+          ? 'npm ci'
+          : 'npm install';
+
+      default:
+        throw new Error(`unexpected package manager ${this.packageManager}`);
+    }
   }
 
   private loadDependencies(outdir: string) {
@@ -1236,10 +1286,10 @@ export class NodeBuildWorkflow extends GithubWorkflow {
         },
 
         // if there are changes, creates a bump commit
-        ...options.bump ? [{ run: 'yarn bump' }] : [],
+        ...options.bump ? [{ run: `${project.runScriptCommand} bump` }] : [],
 
         // build (compile + test)
-        { run: 'yarn build' },
+        { run: `${project.runScriptCommand} build` },
 
         // anti-tamper check (fails if there were changes to committed files)
         // this will identify any non-commited files generated during build (e.g. test snapshots)